+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Remove usage of PHP 7.3 constants JSON_THROW_ON_ERROR
+
 ## [1.3.1] - 2020-11-03
 ### Changed
 - Improve master form styles