+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove usage of PHP 7.3 constants JSON_THROW_ON_ERROR
 - Fix typo in snapshot form ajax logic
 
+### Removed
+- Remove unused file
+
 ## [1.3.1] - 2020-11-03
 ### Changed
 - Improve master form styles